move tests to indexing app
